Break up files into directories and subfiles
